pyproject.toml
[project]
name = "text-to-insight-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.66.0",
    "langchain>=1.2.10",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
